feat(v6.1): title consistency, Events tier selector + full local review pass (rc13)
Title-style consistency + a window-independent Events tier, then a 7-reviewer
local AI code review over the whole v6.1 branch with the valid findings fixed.

Dashboard
- widgetCard gains an icon chip; Battery/Energy/Config cards now match the
  Power Line-quality / Remote-servers title style.
- Events tier selector (Power / +Diagnostics / All): a window-INDEPENDENT filter
  (eventTierOf/isLifecycleEvent/eventPassesTier) so widening the range still
  surfaces power events; per-type checkboxes become optional narrowing.
- Tooltip clears on refresh; *_INACTIVE events read green; DEGRADED remote shows
  "degraded" not "unreachable"; Overview self-test card only when a test ran.

Tiers / events
- TUI: BATTERY_HEALTH_WARNING/CRITICAL + BATTERY_REPLACEMENT_PREDICTED are now
  Power-tier (were hidden in Diagnostics at default verbosity).
- Dashboard tier-1 patterns fixed: emitted name is BATTERY_LOW (not LOW_BATTERY);
  battery-health events are tier-1.

Local-review fixes
- reports.py: battery-health confidence read from `detail` (was always 0%).
- shutdown/plan.py + api.py: redact raw shutdown commands for anonymous readers.
- monitor.py + nut_control.py + api.py: scheduled self-test shares the per-UPS
  control lock (moved into nut_control); api resolvers unified.
- config.py: critical_score must be < warn_score; explicit-null numerics are a
  config error; self_test/reports schedule/time/day/format strings validated.
- stats.py: purge battery_health/self_tests; pre-existing read methods re-check
  _conn after the lock (+ a corrected race test).
- energy.py format_cost numeric spec; battery history window follows
  min_history_days; replacement re-warns weekly; scheduler annotation;
  reload classification meta-test.

2770 unit tests pass; every file >=95%; node --check, mkdocs --strict, all
example configs validate.

+++ src/eneru/health/prediction.py
@@ -63,7 +63,7 @@ def runtime_score(current_runtime_s: Optional[float],
 
 def capacity_score(runtime_history: List[Tuple[float, float]],
                    nominal_runtime_s: Optional[float],
-                   *, min_history_days: float = 14.0) -> Optional[float]:
+                   *, min_history_days: int = 14) -> Optional[float]:
     """Capacity degradation inferred from the runtime TREND, not instantaneous
     charge (which is state-of-charge, not capacity).
 
@@ -132,7 +132,7 @@ def compute_terms(*, current_runtime_s: Optional[float],
                   battery_install_date: Optional[str],
                   expected_life_years: float,
                   now: float,
-                  min_history_days: float = 14.0) -> Dict[str, Optional[float]]:
+                  min_history_days: int = 14) -> Dict[str, Optional[float]]:
     """Compute all five term sub-scores (each 0-100 or None=unavailable)."""
     return {
         "capacity": capacity_score(runtime_history, nominal_runtime_s,
